Module openal::al

Reexports

pub use self::types::*;

Modules

ffi

listener
Functions for setting and retrieving the properties of the listener. One listener is implied for each context.

types

Structs

Buffer
A reference to a buffer object

Source
A reference to a source object

Enums

DistanceModel

Error
An OpenAL error code

Format

SourceType

Functions

delete_buffers
Delete the buffers and free the resources they use. Buffers that are currently in use by a source cannot be deleted.

delete_sources
Delete the sources.

gen_buffers
Generate a one or more buffer objects.

gen_sources
Generate a one or more source objects.

get_distance_model

get_doppler_factor

get_doppler_velocity

get_error
Returns the current error state then clears it.

get_extensions

get_renderer

get_speed_of_sound

get_vendor

get_version

pause_sources

play_sources

rewind_sources

set_distance_model

set_doppler_factor

set_doppler_velocity

set_speed_of_sound

stop_sources