Struct openal::al::Source

pub struct Source {
    // some fields omitted
}

A reference to a source object

Methods

impl Source

fn gen() -> Source

Generate a single source object.

fn delete(self)

Delete the source.

If an error occurs, the source will not be deleted. The error can be detected using get_error. The source may be deleted if it is currently playing. In this case it will be stopped, then destroyed.

fn play(&self)

Play the buffers attached to the source.

fn is_playing(&self) -> bool

fn pause(&self)

Pause the source.

fn is_paused(&self) -> bool

fn stop(&self)

Stop the source

fn is_stopped(&self) -> bool

fn rewind(&self)

Rewind the source to the initial state.

fn is_initial(&self) -> bool

fn queue_buffer(&self, buffer: &Buffer)

Queue a single buffer on the source.

fn queue_buffers(&self, buffers: &[Buffer])

Queue the buffers on the source to be played in sequence.

fn unqueue_buffer(&self, buffer: &mut Buffer)

Remove a single buffer from the queue.

fn unqueue_buffers(&self, buffers: &mut [Buffer])

Remove a set of buffers from the queue.

fn get_buffers_queued(&self) -> usize

fn get_buffers_processed(&self) -> usize

fn get_pitch(&self) -> ALfloat

fn set_pitch(&self, value: ALfloat)

Set the pitch multiplier (should be positive).

fn get_gain(&self) -> ALfloat

fn set_gain(&self, value: ALfloat)

Set the source gain (should be positive).

fn get_max_distance(&self) -> ALfloat

Used with the Inverse Clamped Distance Model to set the distance where there will no longer be any attenuation of the source.

fn set_max_distance(&self, value: ALfloat)

Set the max distance for the source.

fn get_rolloff_factor(&self) -> ALfloat

fn set_rolloff_factor(&self, value: ALfloat)

Set the rolloff factor for the source (default is 1.0).

fn get_reference_distance(&self) -> ALfloat

The distance under which the volume for the source would normally drop by half (before being influenced by rolloff factor or max distance).

fn set_reference_distance(&self, value: ALfloat)

Set the reference distance.

fn get_min_gain(&self) -> ALfloat

fn set_min_gain(&self, value: ALfloat)

Set the minimum gain for the source.

fn get_max_gain(&self) -> ALfloat

fn set_max_gain(&self, value: ALfloat)

Set the maximum gain for the source.

fn get_cone_outer_gain(&self) -> ALfloat

fn set_cone_outer_gain(&self, value: ALfloat)

Set the gain when outside the oriented cone.

fn get_cone_inner_angle(&self) -> ALfloat

fn set_cone_inner_angle(&self, value: ALfloat)

Set the gain when inside the oriented cone.

fn get_cone_outer_angle(&self) -> ALfloat

fn set_cone_outer_angle(&self, value: ALfloat)

Set the outer angle of the sound cone, in degrees. (default is 360.0)

fn get_position(&self) -> [ALfloat; 3]

fn set_position(&self, values: [ALfloat; 3])

Set the position of the source.

fn get_velocity(&self) -> [ALfloat; 3]

fn set_velocity(&self, values: [ALfloat; 3])

Set the velocity vector of the source.

fn get_direction(&self) -> [ALfloat; 3]

fn set_direction(&self, values: [ALfloat; 3])

Set the direction vector of the source.

fn is_relative(&self) -> bool

fn set_relative(&self, value: bool)

Set whether the positions are relative to the listener. (default is false)

fn get_type(&self) -> SourceType

fn set_type(&self, value: SourceType)

Set the source type.

fn is_looping(&self) -> bool

fn set_looping(&self, value: bool)

Set looping on/off for this source.

fn get_buffer(&self) -> Buffer

fn set_buffer(&self, buffer: Buffer)

Set the attached buffer.

fn get_sec_offset(&self) -> ALfloat

fn set_sec_offset(&self, value: ALfloat)

Set the playback position, expressed in seconds.

fn get_sample_offset(&self) -> ALfloat

fn set_sample_offset(&self, value: ALfloat)

Set the playback position, expressed in samples.

fn get_byte_offset(&self) -> ALfloat

fn set_byte_offset(&self, value: ALfloat)

Set the playback position, expressed in bytes.