Struct openal::al::Buffer

pub struct Buffer {
    // some fields omitted
}

A reference to a buffer object

Methods

impl Buffer

fn gen() -> Buffer

Generate a single buffer object.

fn delete(self)

Delete the buffer and free the resources it uses. Buffers that are currently in use by a source cannot be deleted.

unsafe fn buffer_data<T>(&self, format: Format, data: &[T], freq: ALsizei)

Fill the buffer with PCM audio data.

fn get_frequency(&self) -> ALint

The frequency of the buffer in Hz.

fn get_bits(&self) -> ALint

The bit depth of the buffer.

fn get_channels(&self) -> ALint

The number of channels in the buffer.

fn get_size(&self) -> ALint

The size of the buffer in bytes.