Struct openal::alc::Device

pub struct Device {
    // some fields omitted
}

Methods

impl Device

fn open(devicename: Option<&str>) -> Option<Device>

fn close(self) -> Result<(), Device>

Closes the device.

The device will not be closed if it contains any contexts or buffers. If this is the case, the device will be returned again, wrapped in Err.

fn get_error(&self) -> ALCenum

fn get_string(&self, param: ALCenum) -> String

fn create_context(&self, attr_list: &[ALCint]) -> Option<Context>